+++ src/main/java/likelion/festival/booth/ui/BoothController.java
@@ -3,12 +3,14 @@ package likelion.festival.booth.ui;
 import likelion.festival.booth.application.dto.BoothDayLocationDto;
 import likelion.festival.booth.application.dto.BoothDto;
 import likelion.festival.booth.application.dto.BoothFilterDto;
+import likelion.festival.comment.appliction.dto.CommentRequestDto;
+import likelion.festival.comment.appliction.dto.CommentResponseDto;
 import likelion.festival.dto.*;
 import likelion.festival.booth.domain.Booth;
 import likelion.festival.booth.application.BoothService;
 import likelion.festival.menu.application.dto.MenuRequestDto;
 import likelion.festival.menu.application.dto.MenuResponseDto;
-import likelion.festival.service.CommentService;
+import likelion.festival.comment.appliction.CommentService;
 import likelion.festival.service.ImageService;
 import likelion.festival.service.LikesService;
 import likelion.festival.menu.application.MenuService;
