+++ src/main/java/likelion/festival/comment/domain/Comment.java
@@ -1,8 +1,9 @@
-package likelion.festival.entity;
+package likelion.festival.comment.domain;
 
 import com.fasterxml.jackson.annotation.JsonBackReference;
 import com.sun.istack.NotNull;
 import likelion.festival.booth.domain.Booth;
+import likelion.festival.entity.BaseEntity;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
