+++ src/main/java/likelion/festival/comment/domain/repository/CommentRepository.java
@@ -1,10 +1,9 @@
-package likelion.festival.repository;
+package likelion.festival.comment.domain.repository;
 
-import likelion.festival.entity.Comment;
+import likelion.festival.comment.domain.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
-import java.util.Optional;
 
 public interface CommentRepository extends JpaRepository<Comment, Long> {
     List<Comment> findByBooth_IdAndActiveOrderByCreatedDateTimeDesc(Long id,Boolean active);
